refactor: extract the fence checker into tasks/fences.py
Closes #113. A move, not a rewrite: no behaviour changes, and
`rhiza-task list` hashes identically before and after.

quality.py held two things -- the fenced-example checker and seven unrelated
task bodies -- and had reached 904 lines with a docstring describing only the
second group. Nothing breached the complexity ceiling and `complexity` passed
throughout; the maintainability index falling 44.55 to 36.52 across the two
changes that built the checker up was the signal, and cohesion was the
problem.

    quality.py   904 -> 335 lines, MI 36.52 -> 62.24
    fences.py         631 lines, MI 39.19

One name crosses the boundary, and rule 4 chose the design rather than merely
permitting it. CLAUDE.md forbids an underscore-prefixed name from crossing a
module boundary, so keeping the @task in quality.py while moving the helpers
meant either promoting seven private functions -- turning an internal
decomposition into an API -- or giving the new module a single public entry
point taking the gate's whole body. It has one, `check`, and the nine helpers
stayed private. CLAUDE.md now records that as the shape to reach for the next
time a task body outgrows its module.

The task registration stays in quality.py with the argument for why the gate
exists, because that is where a reader looking for a gate looks; fences.py
carries the argument for how it checks, which is what had no home.

Two things the move surfaced rather than caused:

conftest's `recorder` patches uv per module from a hand-maintained list, so a
module binding `uv_run` without registering a task still has to be named
there. It is, with a comment saying the list tracks subprocess stubbing and
not `@task` presence.

The generated scratch script was called `fences.py`, which now collides with
the module in a traceback or a grep -- it caught out a search during this very
change. Renamed to `result_fences.py`, which also says what it holds.

The suite imports the module as `fence_checker`: `fences` is a local variable
holding a fence list throughout test_tasks.py, and the module shadowed it.

363 tests, coverage 100. Both invariant greps unchanged -- two lines and none.
